所有接口数据接入，个人详情里的大部分接口接入

+++ app/src/main/java/com/yjkj/property_management/entity/OwnerEntity.kt
@@ -6,7 +6,7 @@ import java.io.Serializable
 data class OwnerEntity(
     var createBy: Any?,
     var createTime: String?,
-    var markerId: Int?,
+    var markerId: Long?,
     var markerName: Any?,
     var ownerAccountName: Any?,
     var ownerAddress: String?,
@@ -88,7 +88,7 @@ data class OwnerEntity(
     var remark: Any?,
     var updateBy: Any?,
     var updateTime: Any?,
-    var userId: Int?,
+    var userId: Long?,
     var userType: String?,
     var dangerType : Any?,
     var physiologicalState : String?
